multi-cloud-agent/backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn

authlib
passlib[bcrypt]
python-jose[cryptography]
cryptography
bcrypt==3.2.2
slowapi

sqlalchemy
psycopg2-binary

httpx
requests
python-dotenv
python-multipart
itsdangerous

pydantic
email-validator
pydantic-settings

google-generativeai
google-auth
google-search-results
# annoy

boto3
azure-identity
azure-mgmt-compute
azure-mgmt-storage
google-cloud-compute
google-cloud-storage

Pillow
numpy
selenium
